refactor: moved assertions in cs_assertions.go
